split out ExpandBracesTests
